Stop taking tarballs from Sourceforge
Downloading from Sourceforge fails intermitently and regularly, to the
point that it's just not reliable as a source.

That's where upstream puts their tarballs though.

Fortunately, Fedora mirrors upstream tarballs to build their packages,
so we can take them there.

The hash is the same as the upstream one, proving that Fedora didn't
tamper with the tarballs. (as long as sha256 is not broken of course)

Those URLs aren't nice to work with (especially to update to newer
releases), but at least they work reliably.