Add the dependencies for the bpm plugin
Picard has a system to install and load plugins. However some of those
plugins have dependencies, and Picard won't install them.

The BPM plugin is a very popular one which requires 2 additional Python
modules.

Upstream says this should be the only popular plugin with dependencies,
so it should be ok for us to bundle them.

Hopefully we won't have to bundle much more, as that would eventually
start being harmful to all users who don't use those plugins.

Eventually Picard itself might improve its plugin system to the point it
installs plugins with their dependencies, and we will be able to remove
those modules from the Flatpak build.

The two new modules are named "bpm-plugin-dep-* to clearly identify them
as dependencies of a plugin, not Picard itself, so that it is easier to
remove them when the time comes.

Fixes #35